+++ packages/hl7v2-cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rethinkhealth/hl7v2-cli
 
+## 0.12.0
+
+### Patch Changes
+
+- @rethinkhealth/hl7v2@0.12.0
+- @rethinkhealth/hl7v2-config@0.12.0
+
 ## 0.11.0
 
 ### Patch Changes